% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-documentations.R
\docType{data}
\name{annotations}
\alias{annotations}
\title{Saccharomyce cerevisiae genes annotations.}
\format{A data.table with 4 rows and 5 variables.

\describe{
  \item{chromosome}{Gene's chromosome name}
  \item{start}{Gene's transcript start site}
  \item{end}{Gene's transcript termination site}
  \item{gene_name}{Gene's SGD name}
  \item{strand}{Gene's strand}
}}
\usage{
annotations
}
\description{
A dataset containing a subset of four genes from saccer3 gene annotations.
}
\keyword{datasets}

