% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotationSet.R
\name{as.annotationsSet}
\alias{as.annotationsSet}
\title{Convert a data.table into a geneAnnotation object.}
\usage{
as.annotationsSet(data)
}
\arguments{
\item{data}{A \code{\link[data.table]{data.table}} with 6 columns corresponding to the genes'
chromosome, start posittion, end position, gene name, gene score and gene strand.}
}
\value{
An object of class geneAnnotation with the following components:
       \item{chromosome}{Gene chromosome.}
       \item{start}{Gene transcription start site.}
       \item{end}{Gene transcription termination site.}
       \item{name}{Gene name.}
       \item{score}{Gene score.}
       \item{strand}{Gene strand.}
}
\description{
Create a geneAnnotation object containing the informations on
chromosomes, start and end positions, strand, score and name of genes.
}
\examples{
data(annotations)
as.annotationsSet(annotations)
}

