% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-documentations.R
\docType{data}
\name{introns}
\alias{introns}
\title{Saccharomyce cerevisiae introns annotations.}
\format{A data.table with 337 rows and 5 variables.

\describe{
  \item{chromosome}{Intron' chromosome name}
  \item{start}{Intron's transcript start site}
  \item{end}{Intron's transcript termination site}
  \item{gene_name}{Gene's SGD name}
  \item{strand}{Intron's strand}
}}
\usage{
introns
}
\description{
A dataset containing saccer3 introns annotations.
}
\keyword{datasets}

