% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-documentations.R
\docType{data}
\name{rna_seq_signals}
\alias{rna_seq_signals}
\title{Coverage dataset for FLO8, CDC14 and SNC1 gene.}
\format{An object of type coverageDataSet.}
\usage{
rna_seq_signals
}
\description{
A dataset containing the RNA-seq coverage values
in all samples for FLO8, CDC14 and SNC1 gene.
}
\keyword{datasets}

