% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yCrypticRNAs.R
\docType{package}
\name{yCrypticRNAs}
\alias{yCrypticRNAs}
\alias{yCrypticRNAs-package}
\title{Cryptic transcription analysis in yeast.}
\description{
This package contains functions to calculate cryptic scores for
genes using the ratio (Cheung et al., 2008), the 3' enrichment method (DeGennaro
et al., 2013) and the probabilistic method. It also provide methods to estimates
cryptic transcription start sites.
}

