% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-documentations.R
\docType{data}
\name{yer109c}
\alias{yer109c}
\title{Coverage dataset for FLO8 gene.}
\format{An object of type geneCoverage.}
\usage{
yer109c
}
\description{
A dataset containing the RNA-seq coverage values
in all samples for FLO8 gene.
}
\keyword{datasets}

